Extract Camera Hook
