+++ src/apps/chat/components/composer/CameraCaptureButton.tsx
@@ -1,12 +1,14 @@
 import * as React from 'react';
 
-import { Box, Button, CircularProgress, IconButton, LinearProgress, Modal, ModalClose, Option, Select, Sheet, Slider, Typography } from '@mui/joy';
+import { Box, Button, CircularProgress, IconButton, LinearProgress, Modal, ModalClose, Option, Select, Sheet, Typography } from '@mui/joy';
 import AddAPhotoIcon from '@mui/icons-material/AddAPhoto';
 import DownloadIcon from '@mui/icons-material/Download';
 import InfoIcon from '@mui/icons-material/Info';
 import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
 
 import { hideOnDesktop, hideOnMobile } from '~/common/theme';
+
+import { useCamera } from '~/common/components/useCamera';
 
 const isDevelopment = process.env.NODE_ENV === 'development';
 
@@ -16,74 +18,17 @@ export function CameraCaptureButton(props: { onOCR: (ocrText: string) => void })
   const [availableDevices, setAvailableDevices] = React.useState<MediaDeviceInfo[]>([]);
   const [selectedDeviceIdx, setSelectedDeviceIdx] = React.useState<number | null>(null); // availableDevices[0
   const [open, setOpen] = React.useState(false);
-  const [zoomControl, setZoomControl] = React.useState<React.ReactNode>(null);
   const [ocrProgress, setOCRProgress] = React.useState<number | null>(null);
   const [showInfo, setShowInfo] = React.useState(false);
-  const [infoText, setInfoText] = React.useState<string | null>(null);
-  const videoRef = React.useRef<HTMLVideoElement>(null);
 
   // external state
   // const theme = useTheme();
 
   // device selection
   const selectedDevice = selectedDeviceIdx !== null ? availableDevices[selectedDeviceIdx] ?? null : null;
+  const { startCamera, stopCamera, infoText, zoomControl, videoRef } = useCamera(selectedDevice);
 
-
-  // camera operations
-  const startCamera = React.useCallback(async () => {
-    if (navigator.mediaDevices && navigator.mediaDevices.getUserMedia && selectedDevice !== null) {
-      // setTextDebug(JSON.stringify(navigator.mediaDevices.getSupportedConstraints(), null, 2));
-      const constraints: MediaStreamConstraints & { video: { zoom: boolean } } = {
-        video: {
-          deviceId: selectedDevice.deviceId,
-          width: { ideal: 1920 }, // or any desired width
-          height: { ideal: 1440 }, // or any desired height
-          frameRate: { ideal: 30 }, // or any desired frame rate
-          zoom: true, // added for requesting zooming
-        },
-      };
-      try {
-        const stream: MediaStream = await navigator.mediaDevices.getUserMedia(constraints);
-        if (videoRef.current)
-          videoRef.current.srcObject = stream;
-
-        const [track] = stream.getVideoTracks();
-        if (track) {
-          // Get capabilities (for the zoom ranges)
-          const capabilities = track.getCapabilities() as MediaTrackCapabilities & { zoom: { min: number, max: number, step: number } };
-          const settings = track.getSettings();
-          setInfoText(`Settings: ${JSON.stringify(settings, null, 2)}\n\nCapabilities: ${JSON.stringify(capabilities, null, 2)}`);
-
-          // Map zoom to a slider element.
-          if (capabilities.zoom) {
-            const { min, max, step } = capabilities.zoom;
-            const control = <Box sx={{ display: 'flex', mx: 3 }}><Slider
-              color='neutral'
-              min={min} max={max} step={step} defaultValue={1}
-              onChange={(_event, value) => track.applyConstraints({ advanced: [{ zoom: value as number }] } as any)}
-            /></Box>;
-            setZoomControl(control);
-          }
-        }
-
-      } catch (error) {
-        console.error('Error accessing camera:', error);
-      }
-    }
-  }, [selectedDevice]);
-
-  const stopCamera = React.useCallback(() => {
-    if (videoRef.current && videoRef.current.srcObject) {
-      const stream = videoRef.current.srcObject as MediaStream;
-      const tracks = stream.getTracks();
-      tracks.forEach((track) => track.stop());
-      videoRef.current.srcObject = null;
-    }
-  }, []);
-
-
-  // effect to close/start camera on configuration changes
-
+  // close/stream camera on configuration changes
   React.useEffect(() => {
     if (open) startCamera().then(() => null);
     else stopCamera();
@@ -93,8 +38,7 @@ export function CameraCaptureButton(props: { onOCR: (ocrText: string) => void })
   }, [open, startCamera, stopCamera]);
 
 
-  // effect to enumerate devices and auto-select the back camera (or the first camera)
-
+  // enumerate devices and auto-select the back camera (or the first camera)
   React.useEffect(() => {
     // do this only once, ever
     if (navigator.mediaDevices)
@@ -121,7 +65,6 @@ export function CameraCaptureButton(props: { onOCR: (ocrText: string) => void })
   };
 
   const handleDeviceSelected = (_event: any, value: number | null) => setSelectedDeviceIdx(value);
-
 
   const renderVideoFrameToCanvas = (): HTMLCanvasElement => {
     // paint the video on a canvas, to save it
@@ -174,7 +117,7 @@ export function CameraCaptureButton(props: { onOCR: (ocrText: string) => void })
     {isDevelopment && <Button
       fullWidth variant='plain' color='neutral' onClick={handleCameraButtonClicked} startDecorator={<AddAPhotoIcon />}
       sx={{ ...hideOnMobile, justifyContent: 'flex-start' }}>
-      Recognize
+      OCR
     </Button>}
 
     {/* The actual Capture Dialog */}
